Created ops file; Moved isString function to ops
