--- runtime/isString.ts
@@ -1,3 +1,0 @@
-export function isString(x: any): boolean {
-    return x != undefined && typeof x.valueOf() === 'string';
-}
